[Slider] Respect direction where focus comes from when focusing thumb
Previously, the focused thumb index was retained when visiting the same
slider multiple times. Now the focused thumb will be the one on the left
when coming from left and the one on the right when coming from right.
The focused thumb isn't changed when coming from above or below.